requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyngrok==6.0.0
python-dotenv==1.0.0
pytube==15.0.0
PyYAML==6.0
requests==2.31.0
SQLAlchemy==2.0.17
stripe==5.4.0
typing_extensions==4.7.0
urllib3==2.0.3
Werkzeug==2.3.6
